requirements.txt
gymnasium<1.0.0
jax-jumpy==1.0.0 ; python_version >= "3.8" and python_version < "3.11"
matplotlib
moviepy
numpy<2.0
pandas
protobuf
pygame
stable-baselines3
tqdm
wandb
torchrl
tensordict
tyro
python-dotenv
